Sanitize sound source creation
Now sound sources are created even when no attached buffer is available,
so that the program can 'use' them without crashing. This is largely
done by enforcing the use of the nothrow version of the new operator
for SoundSource allocation (which was actually wrongly assumed before!)

The Audio::system() call has also been simplified, letting any exception
in the audio system allocation bleed through to the caller.